// type.ts
export type Pool = {
  id: string;
  coinName: string;
  symbol: string;
  sCoinType: string | null;
  coinType: string;
  marketCoinType: string;
  coinDecimal: number;
  coinPrice: number;
  borrowApr: number;
  borrowApy: number;
  supplyApr: number;
  supplyApy: number;
  supplyCoin: number;
  borrowCoin: number;
  reserveCoin: number;
  utilizationRate: number;
  highKink: number;
  midKink: number;
  reserveFactor: number;
  borrowWeight: number;
  borrowFee: number;
  marketCoinSupplyAmount: number;
  minBorrowAmount: number;
  baseBorrowApr: number;
  baseBorrowApy: number;
  borrowAprOnHighKink: number;
  borrowApyOnHighKink: number;
  borrowAprOnMidKink: number;
  borrowApyOnMidKink: number;
  maxBorrowApy: number;
  borrowIndex: number;
  growthInterest: number;
  supplyAmount: number;
  borrowAmount: number;
  reserveAmount: number;
  conversionRate: number;
  maxSupplyCoin: number;
  maxBorrowCoin: number;
  isIsolated: boolean;
  marketId: string;
  createdAt: string;
  updatedAt: string;
};